:art: Highlight currently displayed banner rows
